+++ src/main/java/edu/hw3/Task4/Task4.java
@@ -1,5 +1,6 @@
-package edu.hw3;
+package edu.hw3.Task4;
 
+import edu.hw3.Util;
 import java.util.List;
 
 public class Task4 {
